Create The_ego_stone
There are 6 Infinity Stones in the Marvel Universe - SoulStone,TimeStone,SpaceStone,MindStone,RealityStone,PowerStone.
It is said that there is a 7th Infinity Stone - The Ego Stone that doesn't appear in Marvel Cinematic Universe (MCU). Thanos is a really ambitious person and has set his eyes on it but in order to obtain that stone he needs to solve a problem and you have to help him out.

Thanos needs to find an InfinityMatrix of order N. A matrix that has Sum of each Row = Sum of each Column = Sum of each diagonal = S (a constant), is called as an InfinityMatrix by Thanos. The Infinity Matrix has all distinct elements in the range 1 and n2.

For any NXN matrix obtain the sum S.

Input:
First line will contain T, number of testcases. Then the testcases follow.
Each testcase contains of a single line of input,one integer N which is the order of the matrix.
Output:
For each testcase, output in a single line S the sum