Create Chef_and_difficult_contests.py
You are given two positive integers a and b. You may apply the following magical operation any number of times (including zero):

choose an arbitrary positive integer d
perform exactly one of the following actions:
add d to a, add d−1 to b
add d−1 to a, add d to b
Chef is asking you to find out if it is possible to obtain a=b. Help him!

Input
The first line of the input contains a single integer T denoting the number of test cases. The description of T test cases follows.
The first and only line of each test case contains two space-separated integers a and b.
Output
For each test case, print a single line containing the string "YES" if it is possible to obtain a=b or "NO" if it is impossible.


This is a question which is purely based on observation. It falls under the category of the questions which are initially difficult to crack, but once you get to the answer, you're like "Could it have been any easier?" ;).