Create Chef_and_ridges.py
We have a rectangular piece of cardboard with width 1 (its height is not important). We are going to fold it in the following way:

1] pick up the right side of the piece of cardboard and move it to the left until it touches the left side
this way, a smaller piece of cardboard with width 1/2 is created; if we unfolded it, we could see a vertical ridge formed in the middle
2] pick up the left side of this new piece of cardboard and move it to the right until it touches the (new) right side
3] pick up the right side of the resulting piece of cardboard and move it to the left until it touches the left side, etc.

Whenever the cardboard is folded, exactly one of its new sides is a newly formed ridge (there may be more, internal ridges formed, but we do not consider these). Let's denote such a ridge created in the k-th folding by Rk.

In total, we fold the piece of cardboard N times. Afterwards, we unfold it and look at the formed ridges. Let's denote the distance of ridge RN (i.e. the last formed outer ridge) from the left side of the original piece of cardboard by DN. For example, D1=1/2 and D2=1/4.

It is possible to express DN as an irreducible fraction x/y. Find this fraction.

Assume that it is possible to fold the piece of cardboard as many times as we want.

Input
The first and only line of the input contains a single integer T denoting the number of test cases. For each test case, a space and an integer N follows.

Output
Print a single line containing 2T space-separated integers. For the i-th test case (1≤i≤T), the 2i−1-th and 2i-th integer should denote x and y — the position of the last ridge as an irreducible fraction.